requirements.txt
avro-python3==1.9.1